remove appendNewImage method
