+++ src/exif-rotate.js
@@ -1,4 +1,5 @@
-import { init, appendNewImage } from './make-canvas';
+import { default_container_id } from './configs';
+import { init } from './make-canvas';
 
 export default class ExifRotate {
   /**
@@ -8,12 +9,18 @@ export default class ExifRotate {
   static showPreviewImage(files, options = {}) {
     const canvas = document.createElement('canvas');
     const ctx = canvas.getContext('2d');
+
     init(files, canvas, ctx, options)
     .then(() => {
-      appendNewImage(canvas, options);
+      const base_64 = canvas.toDataURL('image/jpeg');
+      const new_img = new Image();
+      new_img.setAttribute('src', base_64);
+      const id = options.container_id ? options.container_id : default_container_id;
+      const container = document.getElementById(id);
+      container.appendChild(new_img);
     })
-    .catch(() => {
-      console.log('not working showPreviewImage');
+    .catch((error) => {
+      console.log(error);
     });
   }
 
@@ -30,8 +37,8 @@ export default class ExifRotate {
     .then(() => {
       callback(canvas.toDataURL('image/jpeg'));
     })
-    .catch(() => {
-      console.log('not working getBase64String');
+    .catch((error) => {
+      console.log(error);
     });
   }
 }
